PsychKinectCore: Compatibility fix for 18.04 LTS -> 14.04 LTS
Use pow() + float cast, instead of powf(), to avoid
the new requirement for GLIBC 2.27, which would make
PsychKinectCore fail to load on both Ubuntu 16.04 and 14.04.

All other mex files link and run afaics. However, untested
are Datapixx.mex, Eyelink.mex PsychOpenHMDCore, PsychCV
and the text renderer plugin (not rebuilt), so there may
be more surprises.

+++ PsychSourceGL/Source/Common/PsychKinect/PsychKinect.c
@@ -385,7 +385,7 @@ void PsychKNInit(void) {
 	// Build depths gamma table:
 	for (i=0; i<2048; i++) {
 		v = (float) i/2048.0f;
-		v = powf(v, 3.0f)* 6.0f;
+		v = (float) (pow(v, 3.0) * 6.0);
 		t_gamma[i] = (unsigned short) (v*6*256);
 	}
 }
